requirements.txt
torch
torchvision
transformers
streamlit
matplotlib
numpy
scikit-learn
opencv-python
seaborn
Pillow
